core: always use gulong to store signal handler ids
We inconsistently use gulong,guint,int types to store signal handler
id, but the type returned by g_signal_connect() is a gulong.

This has no practical consequences because a int/guint is enough to
store the value, however it is better to use a consistent type, also
because nm_clear_g_signal_handler() accepts a pointer to the signal id
and thus it must be always called with the same pointer type.

+++ src/devices/nm-device-ethernet.c
@@ -71,8 +71,8 @@ typedef struct Supplicant {
 	NMSupplicantInterface *iface;
 
 	/* signal handler ids */
-	guint iface_error_id;
-	guint iface_state_id;
+	gulong iface_error_id;
+	gulong iface_state_id;
 
 	/* Timeouts and idles */
 	guint iface_con_error_cb_id;
@@ -117,7 +117,7 @@ typedef struct {
 	/* DCB */
 	DcbWait       dcb_wait;
 	guint         dcb_timeout_id;
-	guint         dcb_carrier_id;
+	gulong        dcb_carrier_id;
 } NMDeviceEthernetPrivate;
 
 enum {
